+++ packages/xchain-ethereum/src/client.ts
@@ -10,10 +10,7 @@ import {
   ExplorerUrl,
   TxOverrides,
   GasPrices,
-  GasLimits,
   FeesParams,
-  GasLimitParams,
-  GasLimitsParams,
   FeesWithGasPricesAndLimits,
   InfuraCreds,
 } from './types'
@@ -56,9 +53,14 @@ import {
 export interface EthereumClient {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   call<T>(asset: Address, abi: ethers.ContractInterface, func: string, params: Array<any>): Promise<T>
-
-  estimateGasLimit(params: GasLimitParams): Promise<BigNumber>
-  estimateGasLimits(params: GasLimitsParams): Promise<GasLimits>
+  estimateCall(
+    asset: Address,
+    abi: ethers.ContractInterface,
+    func: string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    params: Array<any>,
+  ): Promise<BigNumber>
+  estimateGasLimit(params: FeesParams): Promise<BigNumber>
   estimateFeesWithGasPricesAndLimits(params: FeesParams): Promise<FeesWithGasPricesAndLimits>
 
   isApproved(spender: Address, sender: Address, amount: BaseAmount): Promise<boolean>
@@ -486,6 +488,31 @@ export default class Client implements XChainClient, EthereumClient {
   }
 
   /**
+   * Call a contract function.
+   * @param {Address} address The contract address.
+   * @param {ContractInterface} abi The contract ABI json.
+   * @param {string} func The function to be called.
+   * @param {Array<any>} params The parameters of the function.
+   * @returns {BigNumber} The result of the contract function call.
+   *
+   * @throws {"address must be provided"}
+   * Thrown if the given contract address is empty.
+   */
+  estimateCall = async (
+    address: Address,
+    abi: ethers.ContractInterface,
+    func: string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    params: Array<any>,
+  ): Promise<BigNumber> => {
+    if (!address) {
+      return Promise.reject(new Error('address must be provided'))
+    }
+    const contract = new ethers.Contract(address, abi, this.provider).connect(this.getWallet())
+    return contract.estimateGas[func](...params)
+  }
+
+  /**
    * Check allowance.
    *
    * @param {Address} spender The spender address.
@@ -495,7 +522,7 @@ export default class Client implements XChainClient, EthereumClient {
    */
   isApproved = async (spender: Address, sender: Address, amount: BaseAmount): Promise<boolean> => {
     try {
-      const txAmount = BigNumber.from(amount.amount().toString())
+      const txAmount = BigNumber.from(amount.amount().toFixed())
       const allowance = await this.call<BigNumberish>(sender, erc20ABI, 'allowance', [this.getAddress(), spender])
       return txAmount.lte(allowance)
     } catch (error) {
@@ -513,7 +540,7 @@ export default class Client implements XChainClient, EthereumClient {
    */
   approve = async (spender: Address, sender: Address, amount?: BaseAmount): Promise<TransactionResponse> => {
     try {
-      const txAmount = amount ? BigNumber.from(amount.amount().toString()) : MAX_APPROVAL
+      const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
       const txResult = await this.call<TransactionResponse>(sender, erc20ABI, 'approve', [
         spender,
         txAmount,
@@ -555,7 +582,7 @@ export default class Client implements XChainClient, EthereumClient {
     gasLimit?: BigNumber
   }): Promise<TxHash> => {
     try {
-      const txAmount = BigNumber.from(amount.amount().toString())
+      const txAmount = BigNumber.from(amount.amount().toFixed())
 
       let assetAddress
       if (asset && assetToString(asset) !== assetToString(AssetETH)) {
@@ -570,7 +597,7 @@ export default class Client implements XChainClient, EthereumClient {
 
       let overrides: TxOverrides = {
         gasLimit: gasLimit || defaultGasLimit,
-        gasPrice: gasPrice && BigNumber.from(gasPrice.amount().toString()),
+        gasPrice: gasPrice && BigNumber.from(gasPrice.amount().toFixed()),
       }
 
       // override `overrides` if `feeOptionKey` is provided
@@ -578,13 +605,11 @@ export default class Client implements XChainClient, EthereumClient {
         const gasPrice = await this.estimateGasPrices()
           .then((prices) => prices[feeOptionKey])
           .catch(() => getDefaultGasPrices()[feeOptionKey])
-        const gasLimit = await this.estimateGasLimit({ asset, recipient, amount, memo, gasPrice }).catch(
-          () => defaultGasLimit,
-        )
+        const gasLimit = await this.estimateGasLimit({ asset, recipient, amount, memo }).catch(() => defaultGasLimit)
 
         overrides = {
           gasLimit,
-          gasPrice: BigNumber.from(gasPrice.amount().toString()),
+          gasPrice: BigNumber.from(gasPrice.amount().toFixed()),
         }
       }
 
@@ -645,16 +670,14 @@ export default class Client implements XChainClient, EthereumClient {
   /**
    * Estimate gas.
    *
-   * @param {EstimateGasOpts} params The transaction options.
+   * @param {FeesParams} params The transaction options.
    * @returns {BaseAmount} The estimated gas fee.
    *
    * @throws {"Failed to estimate gas limit"} Thrown if failed to estimate gas limit.
    */
-  estimateGasLimit = async ({ asset, recipient, amount, memo, gasPrice }: GasLimitParams): Promise<BigNumber> => {
+  estimateGasLimit = async ({ asset, recipient, amount, memo }: FeesParams): Promise<BigNumber> => {
     try {
-      const txAmount = BigNumber.from(amount.amount().toString())
-      // Gas price `Wei` as BigNumber
-      const gasPriceBN = BigNumber.from(gasPrice.amount().toString())
+      const txAmount = BigNumber.from(amount.amount().toFixed())
 
       let assetAddress
       if (asset && assetToString(asset) !== assetToString(AssetETH)) {
@@ -669,7 +692,6 @@ export default class Client implements XChainClient, EthereumClient {
 
         estimate = await contract.estimateGas.transfer(recipient, txAmount, {
           from: this.getAddress(),
-          gasPrice: gasPriceBN,
         })
       } else {
         // ETH gas estimate
@@ -677,7 +699,6 @@ export default class Client implements XChainClient, EthereumClient {
           from: this.getAddress(),
           to: recipient,
           value: txAmount,
-          gasPrice: gasPriceBN,
           data: memo ? toUtf8Bytes(memo) : undefined,
         }
 
@@ -688,26 +709,6 @@ export default class Client implements XChainClient, EthereumClient {
     } catch (error) {
       return Promise.reject(new Error(`Failed to estimate gas limit: ${error.msg ?? error.toString()}`))
     }
-  }
-
-  /**
-   * Estimate gas limits (average, fast fastest).
-   *
-   * @param {GasLimitsParams} params
-   * @returns {GasLimits} The estimated gas limits.
-   */
-  estimateGasLimits = async (params: GasLimitsParams): Promise<GasLimits> => {
-    const { gasPrices, ...otherParams } = params
-    const { fast, fastest, average } = gasPrices
-    return Promise.all([
-      this.estimateGasLimit({ ...otherParams, gasPrice: fast }),
-      this.estimateGasLimit({ ...otherParams, gasPrice: fastest }),
-      this.estimateGasLimit({ ...otherParams, gasPrice: average }),
-    ]).then(([fast, fastest, average]) => ({
-      average,
-      fast,
-      fastest,
-    }))
   }
 
   /**
@@ -725,25 +726,22 @@ export default class Client implements XChainClient, EthereumClient {
       const { fast: fastGP, fastest: fastestGP, average: averageGP } = gasPrices
 
       // gas limits
-      const gasLimits = await this.estimateGasLimits({
+      const gasLimit = await this.estimateGasLimit({
         asset: params.asset,
         amount: params.amount,
         recipient: params.recipient,
         memo: params.memo,
-        gasPrices,
       })
-
-      const { fast: fastGL, fastest: fastestGL, average: averageGL } = gasLimits
 
       return {
         gasPrices,
-        gasLimits,
         fees: {
           type: 'byte',
-          average: getFee({ gasPrice: averageGP, gasLimit: averageGL }),
-          fast: getFee({ gasPrice: fastGP, gasLimit: fastGL }),
-          fastest: getFee({ gasPrice: fastestGP, gasLimit: fastestGL }),
+          average: getFee({ gasPrice: averageGP, gasLimit }),
+          fast: getFee({ gasPrice: fastGP, gasLimit }),
+          fastest: getFee({ gasPrice: fastestGP, gasLimit }),
         },
+        gasLimit,
       }
     } catch (error) {
       return Promise.reject(