+++ packages/xchain-bitcoincash/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v.0.7.1 (2021-03-05)
+
+### Update
+
+- Update `getBalance` to include unconfirmed balances.
+
 # v.0.7.0 (2021-03-02)
 
 ### Breaking change